+++ TombEngine/Scripting/Internal/TEN/Strings/StringsHandler.cpp
@@ -132,7 +132,7 @@ void StringsHandler::ProcessDisplayStrings(float deltaTime)
 				if (str._flags[(size_t)DisplayStringOptions::VerticalCenter])
 					flags |= (int)PrintStringFlags::VerticalCenter;
 
-				m_callbackDrawSring(cstr, str._color, str._position, str._scale, flags);
+				m_callbackDrawSring(cstr, str._color, str._position, str._area, str._scale, flags);
 
 				str._timeRemaining -= deltaTime;
 			}